refactor(speaker): modularize speaker_detection into speaker_detectors (#269)
Extract NER/heuristic logic into focused submodules per RFC-059 while
keeping providers.ml.speaker_detection as a patch-compatible facade.
